refactor(storage): #58 Tier 2.1A — extract record dataclasses to storage/models/
40 record dataclasses (~610 LOC) moved out of src/storage/sqlite.py
(9,559 → 8,994 LOC) into src/storage/models/, grouped by domain:

  storage/models/
    __init__.py     re-exports every record (single source of truth)
    calendar.py     4 records — events, indicators, aliases, vintages
    market.py       4 records — prices, instruments, symbol-history, bars
    indicator.py   10 records — obs/release-schedule/concept-map/fed-comm
    news.py         2 records — articles + trend topics
    analytical.py   4 records — regime / generated note / observation /
                                research artifact
    trading.py      5 records — signals + decision log + position +
                                performance + trading artifact
    messaging.py    6 records — client / conversation / delivery / group
    documents.py    5 records — document + blob + extra + source /
                                release-family registry

storage/sqlite.py re-exports every record from storage.models, so
existing ``from storage.sqlite import XRecord`` consumers (across src/,
tests/, scripts/) keep working without edits.

Also adds ``__init__.py`` files to tests/, tests/calendar/,
tests/calendar/{scheduler,ecb,eodhd}/, tests/oecd/, tests/validation/ —
required because Tier 1.2 split files share basenames across subdirs
(``test_projector.py`` exists under both calendar/ecb/ and
calendar/eodhd/), which pytest's default test-discovery mode rejects
without explicit packaging.

Verification: 1,669 unit tests pass on branch after the extract (same
pattern as before — only network-bound integration suites fail, and
those fail on master too).
