+++ src/storage/models/analytical.py
@@ -1,0 +1,59 @@
+"""Storage records — analytical artifacts (regime snapshots, generated notes, observations, research).
+
+Extracted out of src/storage/sqlite.py as part of issue #58 Tier 2.1A —
+pure mechanical split, no behavior change. The records are re-exported by
+storage.sqlite for backwards compatibility, so existing
+``from storage.sqlite import XRecord`` consumers keep working.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any
+
+
+@dataclass(frozen=True)
+class RegimeSnapshotRecord:
+    snapshot_id: int
+    timestamp: str
+    regime_json: dict[str, Any]
+    trigger_event: str
+    summary: str
+
+
+@dataclass(frozen=True)
+class GeneratedNoteRecord:
+    note_id: int
+    created_at: str
+    note_type: str
+    title: str
+    summary: str
+    body_markdown: str
+    regime_json: dict[str, Any] | None
+    metadata: dict[str, Any]
+
+
+@dataclass(frozen=True)
+class AnalyticalObservationRecord:
+    observation_id: int
+    observation_type: str
+    summary: str
+    detail: str
+    source_kind: str
+    source_id: int
+    created_at: str
+    metadata: dict[str, Any] = field(default_factory=dict)
+
+
+@dataclass(frozen=True)
+class ResearchArtifactRecord:
+    artifact_id: int
+    artifact_type: str
+    title: str
+    summary: str
+    content_markdown: str
+    source_kind: str
+    source_id: int
+    created_at: str
+    tags: list[str] = field(default_factory=list)
+    metadata: dict[str, Any] = field(default_factory=dict)
