+++ src/storage/sqlite.py
@@ -326,617 +326,52 @@ def _add_calendar_country_filter(
     params.append(country_code)
 
 
-@dataclass(frozen=True)
-class StoredEventRecord:
-    source: str
-    event_id: str
-    timestamp: int
-    country: str
-    indicator: str
-    category: str
-    importance: str
-    actual: str | None = None
-    forecast: str | None = None
-    previous: str | None = None
-    revised_previous: str | None = None
-    surprise: float | None = None
-    currency: str = ""
-    unit: str = ""
-    raw_json: dict[str, Any] = field(default_factory=dict)
-    indicator_id: str | None = None
-    event_time_utc: str = ""
-    event_time_precision: str = "datetime"
 
-
-@dataclass(frozen=True)
-class CalendarIndicatorRecord:
-    indicator_id: str               # 'us.inflation.cpi_mom'
-    canonical_name: str             # 'CPI MoM'
-    topic: str
-    country_code: str
-    frequency: str                  # monthly, quarterly, etc.
-    unit: str
-    obs_family_id: str | None = None
-    is_active: bool = True
-    created_at: str = ""
-    updated_at: str = ""
-
-
-@dataclass(frozen=True)
-class CalendarIndicatorAliasRecord:
-    alias_normalized: str           # lowercased, stripped
-    indicator_id: str               # FK → calendar_indicator
-    source: str                     # 'investing'|'forexfactory'|'tradingeconomics'
-    country_code: str
-    alias_original: str = ""
-    created_at: str = ""
-
-
-@dataclass(frozen=True)
-class MarketPriceRecord:
-    symbol: str
-    asset_class: str
-    price: float
-    change_pct: float | None
-    timestamp: int
-    name: str = ""
-
-
-@dataclass(frozen=True)
-class MarketInstrumentRecord:
-    instrument_id: str                      # e.g. "US_SPY"
-    primary_ticker: str                     # current trading ticker
-    name: str                               # "SPDR S&P 500 ETF"
-    asset_class: str                        # equity_etf, equity, bond_etf, commodity_etf
-    market: str                             # "United States equity market"
-    exchange_code: str = ""                 # e.g. "NYSEARCA", "NASDAQ"
-    currency: str = "USD"
-    isin: str = ""
-    openfigi: str = ""
-    composite_figi: str = ""
-    share_class_figi: str = ""
-    cusip: str = ""
-    lei: str = ""
-    primary_provider: str = "tiingo"
-    provider_symbols_json: dict[str, str] = field(default_factory=dict)
-    history_status: str = "provider_continuous"  # provider_continuous|break_detected|stitched|manual_review
-    description_for_agent: str = ""
-
-
-@dataclass(frozen=True)
-class MarketSymbolHistoryRecord:
-    segment_id: str                         # stable, e.g. f"{instrument_id}:{valid_from}:{ticker}"
-    instrument_id: str
-    ticker: str
-    provider_name: str
-    valid_from: str                         # YYYY-MM-DD
-    valid_to: str = ""                      # YYYY-MM-DD or "" for open-ended
-    exchange_code: str = ""
-    isin: str = ""
-    figi: str = ""
-    event_type: str = "listing_start"       # listing_start|ticker_rename|exchange_change|delisting|manual_link
-    mapping_confidence: str = "provider_native"  # provider_native|auto_isin|auto_figi|name_match|manual
-    source_name: str = ""
-    raw_json: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class MarketPriceBarRecord:
-    instrument_id: str
-    date: str                               # YYYY-MM-DD
-    bar_interval: str                       # "1d"
-    open: float
-    high: float
-    low: float
-    close: float
-    volume: float
-    source_name: str                        # "Tiingo"
-    source_symbol: str                      # ticker used at the provider
-    source_segment_id: str = ""
-    adjusted_open: float | None = None
-    adjusted_high: float | None = None
-    adjusted_low: float | None = None
-    adjusted_close: float | None = None
-    adjusted_volume: float | None = None
-    dividend_cash: float = 0.0
-    split_factor: float = 1.0
-    has_break_detected: bool = False
-    has_pre2018_delisted: bool = False
-    has_missing_corp_acts: bool = False
-    has_mapping_review_needed: bool = False
-    quality_flags_json: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class CentralBankCommunicationRecord:
-    source: str
-    title: str
-    url: str
-    timestamp: int
-    content_type: str
-    speaker: str = ""
-    summary: str = ""
-    full_text: str = ""
-
-
-@dataclass(frozen=True)
-class IndicatorObservationRecord:
-    series_id: str
-    source: str
-    date: str
-    value: float
-    metadata: dict[str, Any] = field(default_factory=dict)
-    obs_family_id: str | None = None
-
-
-@dataclass(frozen=True)
-class IndicatorVintageRecord:
-    series_id: str
-    source: str
-    observation_date: str   # the date being measured
-    vintage_date: str       # when this measurement was published
-    value: float
-    metadata: dict[str, Any] = field(default_factory=dict)
-    obs_family_id: str | None = None
-
-
-@dataclass(frozen=True)
-class CalendarEventVintageRecord:
-    event_id: str
-    provider: str           # connector source ("trading_economics", "ec_bcs", ...)
-    vintage_date: str       # source LastUpdate when available, else observed_at
-    observed_at: str        # authoritative for PIT ordering (ISO-8601 UTC)
-    actual: str | None
-    forecast: str | None
-    previous: str | None
-    metadata: dict[str, Any] = field(default_factory=dict)
-    source_url: str = ""    # snapshotted at insert; revision-aware (issue #36)
-    evidence_archive_url: str | None = None  # Wayback snapshot URL when archived
-
-
-@dataclass(frozen=True)
-class ObsSourceRecord:
-    source_id: str          # 'fred', 'eia', 'treasury_fiscal'
-    source_code: str
-    source_name: str
-    source_type: str        # data_aggregator, government_agency, central_bank, exchange, market_data
-    country_code: str
-    homepage_url: str
-    api_base_url: str
-    is_active: bool
-    created_at: str
-    updated_at: str
-
-
-@dataclass(frozen=True)
-class ObsFamilyRecord:
-    family_id: str                  # 'us.inflation.cpi_all'
-    source_id: str                  # 'fred'
-    provider_series_id: str         # 'CPIAUCSL' (matches indicators.series_id)
-    canonical_name: str             # 'CPI All Urban Consumers'
-    short_name: str
-    unit: str                       # 'index', 'percent', 'billions_usd'
-    frequency: str                  # daily, weekly, monthly, quarterly, annual, irregular
-    seasonal_adjustment: str        # sa, nsa, saar, none
-    country_code: str
-    topic_code: str                 # inflation, employment, rates, energy, fiscal
-    category: str                   # consumer_prices, treasury_yields
-    is_active: bool
-    has_vintages: bool
-    metadata: dict[str, Any] = field(default_factory=dict)
-    created_at: str = ""
-    updated_at: str = ""
-
-
-@dataclass(frozen=True)
-class ObsFamilyDocumentRecord:
-    family_id: str
-    release_family_id: str
-    relationship: str           # produced_by, derived_from, related_to
-    created_at: str
-
-
-@dataclass(frozen=True)
-class ConceptMapRecord:
-    """Cross-source indicator mapping — groups equivalent series under one concept."""
-    concept_id: str             # 'CPI_US', 'GDP_REAL_US'
-    source_id: str              # 'fred', 'bls', 'imf'
-    provider_series_id: str     # 'CPIAUCSL', 'CUUR0000SA0'
-    obs_family_id: str          # FK to obs_family.family_id (may be empty)
-    priority: int = 0           # 1 = authoritative, 2 = secondary, 3 = tertiary
-    role: str = "primary"       # primary, secondary, cross_check
-    notes: str = ""
-    created_at: str = ""
-
-
-@dataclass(frozen=True)
-class ResolvedObservation:
-    """A single resolved value for a concept on a given date, with provenance."""
-    concept_id: str
-    date: str
-    value: float
-    source_id: str
-    provider_series_id: str
-    priority: int
-    role: str
-    alternates: int = 0         # how many other sources also had this date
-    vintage: str = "initial"    # "initial", "revised", or "unknown"
-    revision_count: int = 0     # number of vintage entries for this obs date
-
-
-@dataclass(frozen=True)
-class ReleaseScheduleRecord:
-    concept_id: str           # PK, FK to concept_map
-    rule_type: str            # "day_of_month", "weekday_of_month", "quarter_lag",
-                              # "daily", "weekly", "fixed_dates", "approximate_window"
-    rule_json: dict[str, Any] # type-specific params
-    frequency: str            # "daily", "weekly", "monthly", "quarterly", "annual"
-    release_time_utc: str     # "12:30", "14:00", "" if unknown
-    timezone: str             # "America/New_York", "" if unknown
-    source_authority: str     # "manual", "fred_api", "calendar_events"
-    confidence: str           # "exact", "pattern", "approximate"
-    next_expected: str        # ISO datetime, precomputed
-    last_released: str        # ISO datetime, updated after successful fetch
-    last_checked: str         # ISO datetime, updated on every check
-    is_active: bool = True
-    notes: str = ""
-    created_at: str = ""
-    updated_at: str = ""
-
-
-@dataclass(frozen=True)
-class ReleaseStatusRecord:
-    """Tracks per-release availability status with retry state."""
-    concept_id: str
-    release_date: str           # expected release date (ISO)
-    status: str                 # PENDING, WAITING, FETCHED, CONFIRMED, STALE, FAILED
-    attempt_count: int = 0
-    next_retry: str = ""        # ISO datetime for next retry
-    last_attempt: str = ""      # ISO datetime of last fetch attempt
-    source_used: str = ""       # source_id that provided data
-    data_date: str = ""         # latest observation date actually fetched
-    expected_period: str = ""   # minimum expected observation date
-    provisional: bool = False   # True if using fallback source
-    error: str = ""
-    created_at: str = ""
-    updated_at: str = ""
-
-
-@dataclass(frozen=True)
-class NewsArticleRecord:
-    url_hash: str
-    source_feed: str
-    feed_category: str
-    title: str
-    url: str
-    timestamp: int
-    description: str
-    content_markdown: str
-    impact_level: str
-    finance_category: str
-    confidence: float
-    content_fetched: bool
-    institution: str = ""
-    country: str = ""
-    market: str = ""
-    asset_class: str = ""
-    sector: str = ""
-    document_type: str = ""
-    event_type: str = ""
-    subject: str = ""
-    subject_id: str = ""
-    data_period: str = ""
-    contains_commentary: bool = False
-    language: str = "en"
-    authors: str = ""
-    extraction_provider: str = "keyword"
-
-
-@dataclass(frozen=True)
-class TrendTopicRecord:
-    trend_id: str
-    provider: str
-    provider_topic_id: str
-    title_raw: str
-    topic: str
-    summary: str
-    keywords: list[str] = field(default_factory=list)
-    category: str = ""
-    region: str = "global"
-    popularity_score: float = 0.0
-    provider_rank: int = 0
-    engagement_score: float = 0.0
-    comment_count: int = 0
-    observed_at: int = 0
-    expires_at: int = 0
-    raw_json: dict[str, Any] = field(default_factory=dict)
-    normalized_topic_hash: str = ""
-
-
-@dataclass(frozen=True)
-class RegimeSnapshotRecord:
-    snapshot_id: int
-    timestamp: str
-    regime_json: dict[str, Any]
-    trigger_event: str
-    summary: str
-
-
-@dataclass(frozen=True)
-class GeneratedNoteRecord:
-    note_id: int
-    created_at: str
-    note_type: str
-    title: str
-    summary: str
-    body_markdown: str
-    regime_json: dict[str, Any] | None
-    metadata: dict[str, Any]
-
-
-@dataclass(frozen=True)
-class AnalyticalObservationRecord:
-    observation_id: int
-    observation_type: str
-    summary: str
-    detail: str
-    source_kind: str
-    source_id: int
-    created_at: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class ResearchArtifactRecord:
-    artifact_id: int
-    artifact_type: str
-    title: str
-    summary: str
-    content_markdown: str
-    source_kind: str
-    source_id: int
-    created_at: str
-    tags: list[str] = field(default_factory=list)
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class TradeSignalRecord:
-    signal_id: int
-    signal_type: str
-    title: str
-    summary: str
-    rationale_markdown: str
-    signal: dict[str, Any]
-    confidence: float
-    created_at: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class DecisionLogRecord:
-    decision_id: int
-    decision_type: str
-    title: str
-    summary: str
-    rationale_markdown: str
-    research_artifact_id: int | None
-    signal_id: int | None
-    created_at: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class PositionStateRecord:
-    symbol: str
-    exposure: float
-    direction: str
-    thesis: str
-    updated_at: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class PerformanceRecord:
-    record_id: int
-    metric_name: str
-    metric_value: float
-    period_label: str
-    created_at: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class TradingArtifactRecord:
-    artifact_id: int
-    artifact_type: str
-    title: str
-    summary: str
-    rationale_markdown: str
-    research_artifact_id: int
-    decision_log_id: int | None
-    signal: dict[str, Any]
-    confidence: float
-    created_at: str
-    tags: list[str] = field(default_factory=list)
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class ClientProfileRecord:
-    client_id: str
-    preferred_language: str
-    watchlist_topics: list[str]
-    response_style: str
-    risk_appetite: str
-    investment_horizon: str
-    institution_type: str
-    risk_preference: str
-    asset_focus: list[str]
-    market_focus: list[str]
-    expertise_level: str
-    activity: str
-    current_mood: str
-    emotional_trend: str
-    stress_level: str
-    confidence: str
-    notes: str
-    personal_facts: list[str]
-    last_active_at: str
-    total_interactions: int
-    updated_at: str
-
-
-@dataclass(frozen=True)
-class ConversationMessageRecord:
-    message_id: int
-    client_id: str
-    channel: str
-    thread_id: str
-    role: str
-    content: str
-    created_at: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class DeliveryQueueRecord:
-    delivery_id: int
-    client_id: str
-    channel: str
-    thread_id: str
-    source_type: str
-    source_artifact_id: int | None
-    content_rendered: str
-    status: str
-    delivered_at: str | None
-    client_reaction: str
-    created_at: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class GroupProfileRecord:
-    group_id: str
-    group_name: str
-    group_topic: str
-    group_notes: str
-    member_count: int
-    created_at: str
-    updated_at: str
-
-
-@dataclass(frozen=True)
-class GroupMemberRecord:
-    group_id: str
-    user_id: str
-    display_name: str
-    role_in_group: str
-    personality_notes: str
-    first_seen_at: str
-    last_seen_at: str
-    message_count: int
-
-
-@dataclass(frozen=True)
-class GroupMessageRecord:
-    message_id: int
-    group_id: str
-    thread_id: str
-    user_id: str
-    display_name: str
-    content: str
-    created_at: str
-
-
-@dataclass(frozen=True)
-class DocSourceRecord:
-    source_id: str
-    source_code: str
-    source_name: str
-    source_type: str
-    country_code: str
-    default_language_code: str
-    homepage_url: str
-    is_active: bool
-    created_at: str
-    updated_at: str
-
-
-@dataclass(frozen=True)
-class DocReleaseFamilyRecord:
-    release_family_id: str
-    source_id: str
-    release_code: str
-    release_name: str
-    topic_code: str
-    country_code: str
-    frequency: str
-    default_language_code: str
-    created_at: str
-    updated_at: str
-
-
-@dataclass(frozen=True)
-class DocumentRecord:
-    document_id: str
-    release_family_id: str
-    source_id: str
-    canonical_url: str
-    title: str
-    subtitle: str
-    document_type: str
-    mime_type: str
-    language_code: str
-    country_code: str
-    topic_code: str
-    published_date: str
-    published_at: str
-    status: str
-    version_no: int
-    parent_document_id: str
-    hash_sha256: str
-    created_at: str
-    updated_at: str
-    published_precision: str = ""
-    published_epoch_ms: int = 0
-    created_epoch_ms: int = 0
-    updated_epoch_ms: int = 0
-    # 17-field LLM extraction surface (information-layer port).
-    # All default blank/zero so gov_report / SDMX ingestion paths that
-    # never populate them stay valid.
-    institution: str = ""
-    authors: str = ""
-    data_period: str = ""
-    market: str = ""
-    asset_class: str = ""
-    sector: str = ""
-    event_type: str = ""
-    impact_level: str = ""
-    contains_commentary: bool = False
-    confidence: float = 0.0
-    subject_freetext: str = ""
-
-
-@dataclass(frozen=True)
-class DocumentBlobRecord:
-    document_blob_id: str
-    document_id: str
-    blob_role: str
-    storage_path: str
-    content_text: str
-    content_bytes: bytes | None
-    byte_size: int
-    encoding: str
-    parser_name: str
-    parser_version: str
-    extracted_at: str
-
-
-@dataclass(frozen=True)
-class DocumentExtraRecord:
-    document_id: str
-    extra_json: dict[str, Any]
+# Re-export records extracted in issue #58 Tier 2.1A. Existing
+# ``from storage.sqlite import XRecord`` consumers keep working — the names
+# resolve here exactly as they did before the split.
+from storage.models import (  # noqa: E402
+    AnalyticalObservationRecord,
+    CalendarEventVintageRecord,
+    CalendarIndicatorAliasRecord,
+    CalendarIndicatorRecord,
+    CentralBankCommunicationRecord,
+    ClientProfileRecord,
+    ConceptMapRecord,
+    ConversationMessageRecord,
+    DecisionLogRecord,
+    DeliveryQueueRecord,
+    DocReleaseFamilyRecord,
+    DocSourceRecord,
+    DocumentBlobRecord,
+    DocumentExtraRecord,
+    DocumentRecord,
+    GeneratedNoteRecord,
+    GroupMemberRecord,
+    GroupMessageRecord,
+    GroupProfileRecord,
+    IndicatorObservationRecord,
+    IndicatorVintageRecord,
+    MarketInstrumentRecord,
+    MarketPriceBarRecord,
+    MarketPriceRecord,
+    MarketSymbolHistoryRecord,
+    NewsArticleRecord,
+    ObsFamilyDocumentRecord,
+    ObsFamilyRecord,
+    ObsSourceRecord,
+    PerformanceRecord,
+    PositionStateRecord,
+    RegimeSnapshotRecord,
+    ReleaseScheduleRecord,
+    ReleaseStatusRecord,
+    ResearchArtifactRecord,
+    ResolvedObservation,
+    StoredEventRecord,
+    TradeSignalRecord,
+    TradingArtifactRecord,
+    TrendTopicRecord,
+)
 
 
 def _safe_epoch_ms(value: str | datetime | None) -> int:
